requirements.txt
openai
playwright
beautifulsoup4
load_dotenv